Added CharSet struct to parser, nodes, writer, and runner to handle char ranges and categories
The original code had a very hacky method of using strings as an int32 array
to maintain serialized versions of what is now the CharSet struct.  It exposed
the issue that Go strings are UTF-8 encoded bytes and .NET strings are unicode
char (int32) arrays.  A lot of the tricks used in the original code don't work
when you can't 1:1 map indices in a string to int32 runes.

I'm not sure if it was done to improve performance or reduce memory pressure,
but I'd rather have the cleaner code to start with and optimize any trouble spots.

You can refer to dotnet/corefx src/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexCharClass.cs
for the original code.