use `sync.Pool` for runner pooling
Currently as noted in the comment of `putRunner`, there's no attempt
being made to limit the size of the runner pooling - this can result in
the pool containing a lot of runners that were once created in a spur
but will likely not be used anymore. Instead of trying to do gc within
this code, move the pooling to `sync.Pool` which will deallocated objects
in idle and therefore keep the size of the pool as small as possible. For proper
usage of `sync.Pool` there must be a hard limit on the size of `runtrack`,
`runstrack`, `runcrawl` in the case they are too big they will not be
cached, the `initMatch` function has been adjusted to make sure if one
of three slices weren't cached they will be allocated again.

The pool is global and not per-regexp because sync.Pool has a noCopy
struct which conflicts with the the UnmarshalText function of `Regexp`.
This should have the added benefit that if you're executing many
different regexps they will benefit from each other allocated runners.

The motivation for this change is that I'm seeing a lot of memory (~300MiB) being
hold by these runners until the Go program is restarted which feels like
an unoptimal usage of memory, with this change after a spur of these
runners have been created in a small amount of time they are gracefully
deallocated over time and no longer hold memory indefinitely.